+++ beradio/network.py
@@ -4,17 +4,24 @@ from posix import getpid
 from beradio.config import BERadioConfig
 from beradio.protocol import BERadioProtocol1, BERadioProtocol2
 from beradio.snowflake import SnowflakeGenerator
-from beradio.util import PersistentUniqueIdentifier
+from beradio.util import PersistentUniqueIdentifier, human_unique_id
 
 class NetworkIdentifier(PersistentUniqueIdentifier):
     store_class = BERadioConfig
     attribute = 'network_id'
 
-class GatewayIdentifier(PersistentUniqueIdentifier):
+class IdentifierSnowflake(PersistentUniqueIdentifier):
     store_class = BERadioConfig
     attribute = 'gateway_id'
     datacenter = 0
     id_generator = SnowflakeGenerator(datacenter, getpid()).get_next_id
+
+class IdentifierHuman(PersistentUniqueIdentifier):
+    store_class = BERadioConfig
+    attribute = 'gateway_id'
+    id_generator = human_unique_id
+
+GatewayIdentifier = IdentifierHuman
 
 def protocol_factory(version):