+++ CHANGES.rst
@@ -22,6 +22,7 @@ develop
 - libberadio: use "-mcall-prologues" for producing smaller binaries (~400 bytes)
 - improve logging
 - add nanosecond timestamp to json mqtt message
+- use shorter unique id as gateway id
 
 
 2015-10-27 0.4.4